Death poses, heavy combine
Death poses are restored.
NPCs ragdolls have last hit decals correctly applied to them.
Small chance to spawn heavy combine soldiers.
Poison zombie hull widened to match the model.
Barnacles can drop items when killed.

+++ game/server/globals.cpp
@@ -23,5 +23,5 @@
 #include "tier0/memdbgon.h"
 
 Vector			g_vecAttackDir;
-float				g_iSkillLevel;
+int				g_iSkillLevel;
 bool			g_fGameOver;
